+++ src/Extensions/Trackback.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Nexendrie\Rss\Extensions;
 
+use Nexendrie\Rss\Extensions\ElementTypes\Url;
 use Nexendrie\Rss\Generator;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
@@ -25,7 +26,7 @@ final class Trackback extends BaseExtension
     protected function getElementTypes(): array
     {
         return [
-            self::ELEMENT_ABOUT => "url",
+            self::ELEMENT_ABOUT => "string[]",
             self::ELEMENT_PING => "url",
         ];
     }
@@ -33,5 +34,9 @@ final class Trackback extends BaseExtension
     public function configureItemOptions(OptionsResolver $resolver, Generator $generator): void
     {
         $this->registerElements($resolver);
+        $resolver->setAllowedValues(
+            $this->getElementName(self::ELEMENT_ABOUT),
+            static fn(array $value): bool => array_all($value, static fn (string $value): bool => (new Url())->getValidator()($value))
+        );
     }
 }